Finalized initial implementation for working code
Added and tested all needed logic which downloads, unzips, copies and
links the library to the project. All of these have been implemented
within the AndroidBuildScriptModifier.groovy,
AndroidOpenCVGradlePlugin.java, FilesManager.java,
DownloadManager.java. A dependency was added in build.gradle to provide
the download functionality in a simple and straightforward way.

Modified the createClasspathManifest task to exclude any gradle and
groovy jars only.

The repo is ready for publication albeit more improvements to
implementation is in progress.