Major bug solved!
Used the Gradle Tooling API to compile the opencv AARs using
build-cache-scripts and adds them as dependencies dynamically in the
afterEvaluate block in AndroidOpenCVGradlePlugin.java. Removed the
sourceSets modification block in AndroidBuildScriptModifier.groovy as
it is no longer needed and only copied the native directories in
AndroidOpenCVGradlePlugin.java.

Re-implemented recursiveCopy in FilesManager.java to delete directories
in destination to prevent DirectoryNotEmptyException. Also flipped the
output of checkOrCreateDirectory in FilesManager.java to make more
sensible with method name.