Faced a new issue when creating "clean" tasks
The issue is that there is a case when the md5 cache is not matching
which throws an exception and prevents a clean task from running since
the exception is always thrown in configuration phase of the gradle
build before reaching the task in the execution phase.

Added todo's to convert "execution" logic in
AndroidOpenCVGradlePlugin.java into separate Gradle tasks.

Added more tests with AndroidPluginTest.java and
PluginConfigurationTest.java to ensure new logic being from getting
broken.